+++ src/components/switch/index.ts
@@ -1,1 +1,2 @@
 export {WsSwitch} from './ws-switch.js';
+export type {WsSwitchIconTransition} from './ws-switch.js';